fucking with the autoconfig

+++ autoconfig/cpp/create-clang-format.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/bash
-if [ -d .vim-build ] [ ! -f .clang-format ]; then
+if [ -d .vim-build ] && [ ! -f .clang-format ]; then
     cp ~/.config/nvim/autoconfig/cpp/clang-format ./.clang-format
 fi
